refactor(cli): simplified manifest sync
Lots of it was repetition, so it makes sense to separate it into a separate function with boilerplate where a callback does the actual work.
