+++ packages/cli/package.json
@@ -20,6 +20,7 @@
 		"access": "public",
 		"registry": "https://registry.npmjs.org/"
 	},
+	"main": "dist/index.js",
 	"files": [
 		"src/",
 		"dist/",
